% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{incremental_mean_and_var}
\alias{incremental_mean_and_var}
\title{Incremental Mean and Variance}
\usage{
incremental_mean_and_var(
  X,
  last_mean,
  last_variance,
  last_sample_count,
  sample_weight = NULL
)
}
\arguments{
\item{X}{Matrix like}

\item{last_mean}{A number.}

\item{last_variance}{A number.}

\item{last_sample_count}{A number.}

\item{sample_weight}{vector like.}
}
\value{
A list of of updated mean, updated variance and updated sample count
}
\description{
Incremental Mean and Variance
}
